api_platform/api-platorm-backend/common/pyproject.toml
[project]
name = "common"
version = "1.0"
description = ""
authors = [
    {name = "Sunny Wang"}
]
requires-python = ">=3.10,<3.11"

[tool.poetry]
packages = [{ include = "../common" }]

[tool.poetry.dependencies]
PyYAML = "*"
redis = "*"
sqlalchemy = "*"
python-jose = "*"
snowflake-id = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
